+++ multiqc/modules/picard/AlignmentSummaryMetrics.py
@@ -2,7 +2,7 @@
 
 """ MultiQC submodule to parse output from Picard AlignmentSummaryMetrics """
 
-from collections import OrderedDict, defaultdict
+from collections import OrderedDict
 import logging
 import os
 import re